Refactoring the class Authentication
